Add a few things from Konosuba Plus (not fully working)
The Switch version of Konosuba Plus also contains an
Entergram-published VN using the SNR format. I have added a few
insights gained from RE-ing that into the script. However, it seems
that some of its opcodes behave differently from the ones in Kal/Saku,
so for now I have no intentions of fully supporting this with the
current script. This might change in the future.