SNR generation for Saku
Saku's SNR file can now be losslessly de- and recompiled. This took
more effort than I initially expected, because I first had to
reverse-engineer the structure behind the Saku-specific tables
that I previously ignored.

This should open the doors for a potential English mod of Umineko
Saku, if anyone is interested in going through every single
dialogue line and adding in the translation. I may do a PoC of
this soon.